requirements.txt
feedparser
supabase
python-dateutil
thefuzz
python-Levenshtein
beautifulsoup4
requests
python-dotenv
google-genai
